refactor: changed project layout to be more intuitive
