+++ src/services/logging.ts
@@ -1,5 +1,5 @@
 import { OutputChannel, window, workspace, WorkspaceConfiguration } from 'vscode';
-import { EXTENSION_ID } from '../lib/tools';
+import { EXTENSION_ID } from '../util/tools';
 
 export enum LogLevel {
   DEBUG = 0,
