+++ test/unitTest/document-system-provider.test.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { strictEqual } from 'assert';
-import { FortranDocumentSymbolProvider } from '../../src/features/document-symbol-provider';
+import { FortranDocumentSymbolProvider } from '../../src/fallback-features/document-symbol-provider';
 
 suite('Document Symbol Provider', () => {
   let doc: vscode.TextDocument;
